Models david

+++ src/main/java/com/workable_sb/workable/repositories/RolRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.workable_sb.workable.models.Rol;
 
-public interface RolRepository extends JpaRepository<Rol, Short> {
+public interface RolRepository extends JpaRepository<Rol, Integer> {
 
 }
