+++ src/main/java/com/workable_sb/workable/repositories/TipDocumentoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.workable_sb.workable.models.TipDocumento;
 
-public interface TipDocumentoRepository extends JpaRepository<TipDocumento, Short> {
+public interface TipDocumentoRepository extends JpaRepository<TipDocumento, Integer> {
 
 }
